I made a bunch of ANSI complience changes (so that it iwill compile in Linux and Solaris).  We need to fully check these in Windows next. I updated the .pro file to chang it slightly. We may have to build TinyXML directly into mapimg2. TinyXML does not build as a library in Linux, apparently, only in Windows. More on this later. I went ahead and added the necessary files to CVS and updated the project file. There are still several ANSI changes that need to be made, and a few small layout things. After that I will be ready to roll in the changes to add resampling methods. Eventually that will be done, then Bob and I will test, then cross-programmer testing, and finally RELEASE!!!

+++ mapimg/qinfoframe.cpp
@@ -1,4 +1,4 @@
-// $Id: qinfoframe.cpp,v 1.1 2005/01/14 16:18:50 rbuehler Exp $
+// $Id: qinfoframe.cpp,v 1.2 2005/01/27 18:15:16 jtrent Exp $
 
 
 #include "qinfoframe.h"
@@ -606,7 +606,7 @@ void QInfoFrame::getFill()
 {
    RasterInfo inf( info() );
 
-   int maxValue = mapimg::calcFillValue(inf);
+   double maxValue = mapimg::calcFillValue(inf);
 
    if( maxValue != 0 )
       mapTab->fillEdit->setText( QString::number(maxValue + 2, 'f', 6 ) );
